calculus q1 proofs: remove proof function
I'm pretty sure it's impossible to do this. I wanted it to satisfy the "no more
than two proofs per page" requirement on the instructions, but it doesn't seem
to be possible to do a "minimum height" constraint in Typst. If it is, it's
going to require a lot of ugly hacking, so I think I'm just going to resign to
fulfilling that requirement manually ¯\_(ツ)_/¯.