seminar letter: fix old typo
This diff just randomly showed up in my working tree, as if I'm an amnesiac
looking at the results of actions I cannot remember. I have no idea when I did
it or why, but since it's already here, I suppose it is a fix to a document that
has long passed its expiry date and which no one cares about anymore.